requirements.txt
biopython==1.83
requests==2.31.0
tqdm==4.66.1
beautifulsoup4==4.12.2
